Feature(feat-admin-get-users):adding confirmation message alert before delete user

+++ resources/js/components/admin/AllUsers.vue
@@ -66,10 +66,31 @@ export default {
                 });
         },
         deleteUser(id) {
+
+           this.$swal
+        .fire({
+          title: "Are you sure?",
+          text: "You won't be able to revert this!",
+          icon: "warning",
+          showCancelButton: true,
+          confirmButtonColor: "#3085d6",
+          cancelButtonColor: "#d33",
+          confirmButtonText: "Yes, delete it!",
+        })
+        .then((result) => {
+          if (result.value) {
             axios.delete(apiBase + `admin/deleteuser/${id}`).then((response) => {
                 let i = this.users.map((data) => data.id).indexOf(id);
                 this.users.splice(i, 1);
+                 this.$swal.fire(
+                  "Deleted!",
+                  "Product is deleted successfully",
+                  "success"
+                );
             });
+          };
+        });
+        
         },
         getrooms() {
             axios
